Remove unused code from lab
This commit removes unused or obsolete files from the lab directory. These files included:
- fscan/hfiles.go: This file contained functions and types related to file scanning which are no longer used.
- fscan/hfiles_test.go: This was the corresponding test file for hfiles.go. Since hfiles.go is removed, this is also redundant.
- memchk/main.go: This file contained a main function that was not being used.
- new_processors/main.go: This file was another unused main function.

The removal of these files is part of a broader effort to clean up the codebase and remove unused or obsolete code. This not only makes the codebase cleaner and easier to navigate, but also reduces confusion for any new contributors joining the project.